Add automated and manual PHAR-creation
The PHAR file will be created using humbug/box. It will be scoped, so
that it will not interfere with any other autoloading code.

The code in the .travis.yml-file will cause the PHAR file to be created
automatically by the CI-System everytime a new tag is created. The
resulting PHAR file will then be uploaded as build-artifact and can then
be downloaded from the Release-page.

This requires the GITHUB_API_KEY to be set! This has to be done by the
maintainer!

The PHAR will be created using PHP7.2